# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PQRS API Documentation",
        "description": "API para el sistema de gestión de PQRS (Peticiones, Quejas, Reclamos y Sugerencias) con monitoreo y métricas integradas",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/answers": {
            "get": {
                "tags": [
                    "Answers"
                ],
                "summary": "Get all answers",
                "operationId": "2999246631442e53707908082211c0e4",
                "responses": {
                    "200": {
                        "description": "List of answers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "pqrs_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Esta es la respuesta al PQRS"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Create a new answer",
                "operationId": "bc65d0c796e85fc9970e44886244f30f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pqrs_id",
                                    "user_id",
                                    "description"
                                ],
                                "properties": {
                                    "pqrs_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Esta es la respuesta al PQRS"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Answer created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get all cities",
                "operationId": "c822efcc3a583121c365f5be793afce0",
                "responses": {
                    "200": {
                        "description": "List of cities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Medellín"
                                            },
                                            "state_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "MED"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Create a new city",
                "operationId": "92d1aa8744e6a992db8217b8ebe44cda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "state_id",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bogotá"
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "BOG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pqrs": {
            "get": {
                "tags": [
                    "PQRS"
                ],
                "summary": "Get all PQRS",
                "operationId": "df135e3fb34951900e73ab27d37e9b39",
                "responses": {
                    "200": {
                        "description": "List of PQRS",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "guia": {
                                                "type": "string",
                                                "example": "GU123456"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "document": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "3001234567"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Calle 123 #45-67"
                                            },
                                            "cel_phone": {
                                                "type": "string",
                                                "example": "3009876543"
                                            },
                                            "destiny_city_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "pqrs_type_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Descripción del PQRS"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "status_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PQRS"
                ],
                "summary": "Create a new PQRS",
                "operationId": "c9fa2b9551080c8bbfabce0df24407db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guia",
                                    "name",
                                    "document",
                                    "phone",
                                    "address",
                                    "cel_phone",
                                    "destiny_city_id",
                                    "pqrs_type_id",
                                    "description",
                                    "user_id",
                                    "status_id"
                                ],
                                "properties": {
                                    "guia": {
                                        "type": "string",
                                        "example": "GU123456"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "3001234567"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle 123 #45-67"
                                    },
                                    "cel_phone": {
                                        "type": "string",
                                        "example": "3009876543"
                                    },
                                    "destiny_city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pqrs_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del PQRS"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PQRS created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PQRS creada exitosamente."
                                        },
                                        "pqrs": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pqrs/{id}": {
            "put": {
                "tags": [
                    "PQRS"
                ],
                "summary": "Update a PQRS",
                "operationId": "a758a10f6f382655ed69593887b350b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guia",
                                    "name",
                                    "document",
                                    "phone",
                                    "address",
                                    "cel_phone",
                                    "destiny_city_id",
                                    "pqrs_type_id",
                                    "description",
                                    "user_id",
                                    "status_id"
                                ],
                                "properties": {
                                    "guia": {
                                        "type": "string",
                                        "example": "GU123456"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "3001234567"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle 123 #45-67"
                                    },
                                    "cel_phone": {
                                        "type": "string",
                                        "example": "3009876543"
                                    },
                                    "destiny_city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pqrs_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del PQRS"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PQRS updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PQRS actualizada exitosamente."
                                        },
                                        "pqrs": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PQRS not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PQRS"
                ],
                "summary": "Delete a PQRS (soft delete)",
                "operationId": "a0b03a69b746e5113e6eaa8e6b47d269",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PQRS deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PQRS eliminada exitosamente (soft deleted)."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PQRS not found"
                    },
                    "409": {
                        "description": "PQRS already deleted"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "operationId": "6cd48af64dafcf1c85923dba47e841a4",
                "responses": {
                    "200": {
                        "description": "List of roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Admin"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Administrator role"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new role",
                "operationId": "4eb1ce389f63b4dbfcc5b002014636e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Manager"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Manager role description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer <token>",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}